Make `URLExternal` protected
