+++ dmr/internal/routing.py
@@ -1,0 +1,30 @@
+import dataclasses
+from typing import final
+
+from django.urls.resolvers import URLPattern
+
+from dmr.openapi.objects import PathItem
+
+
+@final
+@dataclasses.dataclass(slots=True, frozen=True)
+class URLExternal:
+    """
+    Represents an external URL that was added to the routing of DMR.
+
+    Prefer :func:`external_path` over using this class directly.
+    See :ref:`external-views` for more info.
+
+    .. versionadded:: 0.13.0
+    .. versionchanged:: 0.14.0
+        Moved to internal and made protected.
+
+    """
+
+    url: URLPattern
+    openapi: PathItem | None = dataclasses.field(kw_only=True)
+
+    def get_url_with_metadata(self) -> URLPattern:
+        """Get the url pattern with attached OpenAPI metadata."""
+        self.url.callback.__dmr_external_openapi__ = self.openapi  # type: ignore[attr-defined]
+        return self.url
