+++ docs/pages/deep-dive/public-api.rst
@@ -105,10 +105,9 @@ Routing
 
 .. autofunction:: dmr.routing.path
 
-.. autoclass:: dmr.routing.URLExternal
-  :members:
-
 .. autofunction:: dmr.routing.external_path
+
+.. autofunction:: dmr.routing.routing_path
 
 
 Meta mixins
